fix: HashAlgorithm shape
