+++ .github/workflows/validate-js.yml
@@ -65,6 +65,9 @@ jobs:
         env:
           REVIEWDOG_GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
+      - name: Test JS
+        run: bun test
+
   lint_js:
     name: JS Lint (eslint, prettier)
     runs-on: ubuntu-latest
